Refactored database tests and the harness thereto into three different files.
